MONGOID-5066 Test coverage for multiple where calls with symbol operators
